+++ control/src/cache/reconcile/heads.rs
@@ -8,6 +8,7 @@ impl Store {
         self.cache.options.rewind = self.input.switch[2];
         self.cache.options.quantize_8 = self.input.switch[3];
         self.cache.options.quantize_6 = self.input.switch[4];
+        self.cache.options.random_impulse = self.input.switch[5];
 
         for i in 0..4 {
             self.reconcile_head(i);